requirement.txt
django
djangorestframework
django-ckeditor
django-redis
googletrans==4.0.0-rc1
gunicorn
whitenoise
